Add start_index argument to 'parse' function
Usual case for emulation programs, like sde or gdb, is to pass first
arguments to the emulator, and let the guest application handle latter
arguments. To support this behavior, we introduce an option to choose
index of the first argument to be handled by POPL. By default it is 1,
as it is usually corresponds to the binary name.